Move workaround to Analyzer
