+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/Analyzer.kt
@@ -1,5 +1,6 @@
 package io.gitlab.arturbosch.detekt.core
 
+import io.github.detekt.tooling.api.spec.ProcessingSpec
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.FileProcessListener
 import io.gitlab.arturbosch.detekt.api.Finding
@@ -9,6 +10,9 @@ import io.gitlab.arturbosch.detekt.api.RuleSetId
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
 import io.gitlab.arturbosch.detekt.api.internal.BaseRule
 import io.gitlab.arturbosch.detekt.api.internal.CompilerResources
+import io.gitlab.arturbosch.detekt.api.internal.DisabledAutoCorrectConfig
+import io.gitlab.arturbosch.detekt.api.internal.FailFastConfig
+import io.gitlab.arturbosch.detekt.core.config.DefaultConfig
 import io.gitlab.arturbosch.detekt.core.rules.IdMapping
 import io.gitlab.arturbosch.detekt.core.rules.associateRuleIdsToRuleSetIds
 import io.gitlab.arturbosch.detekt.core.rules.isActive
@@ -24,7 +28,7 @@ internal class Analyzer(
     private val processors: List<FileProcessListener>
 ) {
 
-    private val config: Config = settings.config
+    private val config: Config = settings.spec.workaroundConfiguration(settings.config)
     private val idMapping: IdMapping =
         associateRuleIdsToRuleSetIds(providers.associate { it.ruleSetId to it.instance(config).rules })
 
@@ -125,4 +129,24 @@ internal class Analyzer(
 
 private fun <T, U, R> Sequence<Pair<T, U>>.mapLeft(transform: (T, U) -> R): Sequence<Pair<R, U>> {
     return this.map { (first, second) -> transform(first, second) to second }
+}
+
+internal fun ProcessingSpec.workaroundConfiguration(config: Config): Config = with(configSpec) {
+    var declaredConfig: Config? = when {
+        configPaths.isNotEmpty() -> config
+        resources.isNotEmpty() -> config
+        useDefaultConfig -> config
+        else -> null
+    }
+
+    if (rulesSpec.activateExperimentalRules) {
+        val defaultConfig = DefaultConfig.newInstance()
+        declaredConfig = FailFastConfig(declaredConfig ?: defaultConfig, defaultConfig)
+    }
+
+    if (!rulesSpec.autoCorrect) {
+        declaredConfig = DisabledAutoCorrectConfig(declaredConfig ?: DefaultConfig.newInstance())
+    }
+
+    return declaredConfig ?: DefaultConfig.newInstance()
 }
